Set the manifest values in `Cargo.toml`
I find packages without authors weird and wouldn't use them. Hence I am
adding myself as the author, with an eMail address that is related to
the project for which I am creating this.

I also updated the description to be more useful.
I am not yet ready to call this `1.0`, I would at least need to test a
few apis and add a readme.

As for the license: I just chose what seems to be the default in the
rust ecosystem.